Added new calibration application
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4910 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/test/geometry/GeometryTest.cpp
@@ -141,7 +141,7 @@ BOOST_AUTO_TEST_CASE( TriMeshProfiling ){
 	BOOST_MESSAGE("toIdxMesh float uint16 get direct3 time: " << timer.getTime());
 
 	// now test if the indexed data is the same as the plain
-	const float epsilon = 1e-5;
+	const float epsilon = (float)1e-5;
 	timer.resetAndResume();
 	for(size_t i=0; i<mesh->getSize(); i++){
 		//std::cout << MetricUtil::dist2(imeshf16->getVertex(i, V1), ((*mesh)[i])[0]) << "<" << epsilon << std::endl;
